Fix Capistrano integration regression
84896ef3 broke the capistrano integration, since we were trying to
access Cogy as if it was loaded by Rails, but it weren't.

Unfortunately, we can't read the already provided `Cogy.cogy_endpoint`
inside the capistrano task without some sort of magic. For the time
being, we'll go with the old behavior, which means we have to still
define the cogy endpoint in the Capistrano task too, even though we
already provide it in the Rails configuration.